Factor out user authentication into middleware and
add tests. Also update tests for user/me accordingly.

NOTE:
Handlers rely on a valid, non-nil :user entity in the request map, to
determine whether or not the request is authorised.

Authorisation requires a signed token that has NOT yet expired.
verify-token checks for expiry, and so our middleware automatically
"revokes" users having expired tokens. This is not obvious from the
middleware, and so we must test for this behaviour.

We update the create-token function api to conveniently test expired
tokens, by configuring it with a time in the past.