Add migrations, db fns, specs for timers and spans
All timer functions are restricted to the user that owns the timer. Time
Spans are not directly manipulated because they have no meaning outside
the context of a timer.

DB queries MUST ensure user isolation for all CRUD operations. Users may
not mess with each others' timers or time spans.