Don't throw away tasks data on fetching tasks
The task data being replaced appeared to contribute to flicker. Also, it
might be better to keep existing task objects around to reduce pressure
on memory, when the number of tasks increase. The assoc would probably
have the effect of evicting all current objects and replacing them with
new ones.