Fix generator to fix flaky create-user factory fn
Prior to this fix, the create-user factory could return the same user in
successive calls, if the generated google ID repeated. This actually
caused a build failure in timer-tests, because we create a second user
and create an acl for the user. Create ACL failed a unique constraint,
presumably because the second time create-user was called, it actually
returned an existing user.